--- lib/src/service/tweets/filtering/builder/conclusion.dart
@@ -1,10 +1,0 @@
-// Copyright 2022 Kato Shinya. All rights reserved.
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided the conditions.
-
-import 'filtering_rule_syntax.dart';
-
-abstract class Conclusion implements FilteringRuleSyntax {
-  /// Returns the built filtering rule.
-  String build();
-}
