+++ lib/src/service/tweets/filtering/channel/post_logical_channel.dart
@@ -1,0 +1,11 @@
+// Copyright 2022 Kato Shinya. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+// Project imports:
+import '../syntax/conjunction_required_syntax.dart';
+
+class PostLogicalChannel extends ConjunctionRequiredSyntax {
+  /// Returns the new instance of [PostLogicalChannel].
+  const PostLogicalChannel(super.buffer);
+}
