+++ lib/src/service/tweets/filtering/operator/conjunction/conjunction_required_operator.dart
@@ -2,6 +2,7 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
+// Project imports:
 import '../filtering_rule_operator.dart';
 
 abstract class ConjunctionRequiredOperator extends FilteringRuleOperator {
