+++ lib/src/service/tweets/filtering/operator/filtering_rule_operator.dart
@@ -2,6 +2,7 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
+// Project imports:
 import '../invalid_filtering_rule_exception.dart';
 import 'validation_result.dart';
