+++ lib/src/service/tweets/filtering/operator/logical/or.dart
@@ -2,6 +2,7 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
+// Project imports:
 import 'logical_operator.dart';
 
 class Or extends LogicalOperator {
