+++ lib/src/service/tweets/filtering/operator/standalone/user_bio_name.dart
@@ -2,6 +2,7 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
+// Project imports:
 import '../validation_result.dart';
 import 'standalone_operator.dart';
