+++ lib/twitter_api_v2.dart
@@ -50,7 +50,7 @@ export 'package:twitter_api_v2/src/service/spaces/spaces_service.dart';
 export 'package:twitter_api_v2/src/service/spaces/topic_data.dart';
 export 'package:twitter_api_v2/src/service/spaces/topic_field.dart';
 export 'package:twitter_api_v2/src/service/tweets/exclude_tweet_type.dart';
-export 'package:twitter_api_v2/src/service/tweets/filtering/builder/filtering_rule_builder.dart'
+export 'package:twitter_api_v2/src/service/tweets/filtering/filtering_rule_buffer.dart'
     show newFilteringRule;
 export 'package:twitter_api_v2/src/service/tweets/filtering_rule_data.dart';
 export 'package:twitter_api_v2/src/service/tweets/filtering_rule_meta.dart';
